Refactor right side panel summary into modular components
- Created `src/feature/rightsidepanel/summary/` directory.
- Split summary logic into modular components: `PresetSummaryRow`, `AuOptionSummaryRow`, `ExRMinMaxSummaryRow`, and `VanillaRoleSummaryRow`.
- Updated `RightSidePanelSummary` to compose these new components.
- Maintained performance optimizations using fine-grained Zustand selectors.
- Ensured all labels are translated and support color tags.
- Verified that all unit and E2E tests pass.
- Satisfied Biome linting and formatting requirements.
